change Makefile
