+++ pkg/util/zookeeper_util.go
@@ -12,14 +12,8 @@ package util
 
 import (
 	"container/list"
-	"context"
 	"fmt"
 	"time"
-
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/util/wait"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/pravega/pravega-operator/pkg/apis/pravega/v1alpha1"
 	"github.com/samuel/go-zookeeper/zk"
@@ -30,34 +24,6 @@ const (
 	PravegaPath = "pravega"
 	ZkFinalizer = "cleanUpZookeeper"
 )
-
-// Wait for pods in cluster to be terminated
-func WaitForClusterToTerminate(kubeClient client.Client, p *v1alpha1.PravegaCluster) (err error) {
-	listOptions := &client.ListOptions{
-		LabelSelector: labels.SelectorFromSet(LabelsForPravegaCluster(p)),
-	}
-
-	err = wait.Poll(5*time.Second, 2*time.Minute, func() (done bool, err error) {
-		podList := &corev1.PodList{}
-		err = kubeClient.List(context.TODO(), listOptions, podList)
-		if err != nil {
-			return false, err
-		}
-
-		var names []string
-		for i := range podList.Items {
-			pod := &podList.Items[i]
-			names = append(names, pod.Name)
-		}
-
-		if len(names) != 0 {
-			return false, nil
-		}
-		return true, nil
-	})
-
-	return err
-}
 
 // Delete all znodes related to a specific Pravega cluster
 func DeleteAllZnodes(p *v1alpha1.PravegaCluster) (err error) {
